[LABS-490] Extract coin splitting functionality into its own module
